+++ src/wai/annotations/main/commands/batch_split/__init__.py
@@ -1,0 +1,6 @@
+"""
+Package containing the specification of the 'domains' sub-command.
+"""
+from ._batch_split_main import batch_split_main, perform_batch_split
+from ._BatchSplitOptions import BatchSplitOptions
+from ._help import batch_split_help